# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-Store Inventory Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # STORES
...
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, shipped, received]
        - name: fromStoreId
          in: query
          schema:
            type: integer
        - name: toStoreId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferList"
    post:
      operationId: createTransfer
      tags: [transfers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/approve:
    post:
      operationId: approveTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/reject:
    post:
      operationId: rejectTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/ship:
    post:
      operationId: shipTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipTransferInput"
      responses:
        "200":
          description: Shipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/receive:
    post:
      operationId: receiveTransfer
      tags: [transfers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveTransferInput"
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"

  # PRODUCTION
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, store_manager, production_manager, sales_officer, finance_officer, approver]
        storeId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        storeId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        storeId:
          type: ["integer", "null"]
        isActive:
          type: boolean

    # STORE
...
    ApprovalAction:
      type: object
      properties:
        notes:
          type: ["string", "null"]

    # TRANSFER
    TransferItem:
      type: object
      required: [productId, requestedQty]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        requestedQty:
          type: number
        approvedQty:
          type: ["number", "null"]
        shippedQty:
          type: ["number", "null"]
        receivedQty:
          type: ["number", "null"]
    Transfer:
      type: object
      required: [id, transferNumber, fromStoreId, toStoreId, status, createdAt]
      properties:
        id:
          type: integer
        transferNumber:
          type: string
        fromStoreId:
          type: integer
        fromStoreName:
          type: ["string", "null"]
        toStoreId:
          type: integer
        toStoreName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, shipped, received]
        notes:
          type: ["string", "null"]
        requestedById:
          type: ["integer", "null"]
        approvedById:
          type: ["integer", "null"]
        shippedAt:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransferItem"
        createdAt:
          type: string
    TransferList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TransferInput:
      type: object
      required: [fromStoreId, toStoreId, items]
      properties:
        fromStoreId:
          type: integer
        toStoreId:
          type: integer
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransferItem"
    ShipTransferInput:
      type: object
      required: [items]
      properties:
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            required: [transferItemId, shippedQty]
            properties:
              transferItemId:
                type: integer
              shippedQty:
                type: number
    ReceiveTransferInput:
      type: object
      required: [items]
      properties:
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            required: [transferItemId, receivedQty]
            properties:
              transferItemId:
                type: integer
              receivedQty:
                type: number

    # PRODUCTION
...